refactor command types
