+++ src/components/command/index.ts
@@ -1,0 +1,3 @@
+export * from './Command';
+export * from './command.types';
+export * from './command.util';
